Process review comments
