+++ src/shared/services/settings.service.ts
@@ -1,15 +1,17 @@
 import { Unsubscriber } from '@shared/utils/papi-util';
 import PapiEventEmitter from '@shared/models/papi-event-emitter.model';
 
+type SettingType<T> = T | null;
+
 /** All message subscriptions - emitters that emit an event each time a setting is updated */
-const messageEmitters = new Map<string, PapiEventEmitter<unknown>>();
+const onDidUpdateSettingEmitters = new Map<string, PapiEventEmitter<string>>();
 
 /**
  * Retrieves the value of the specified setting
  * @param key The string id of the setting for which the value is being retrieved
  * @returns The value of the specified setting, parsed to an object. Returns `null` if setting is not present or no value is available
  */
-const getSetting = (key: string): unknown | null => {
+const getSetting = <T>(key: string): SettingType<T> => {
   const settingString = localStorage.getItem(key);
   return settingString !== null ? JSON.parse(settingString) : null;
 };
@@ -19,9 +21,9 @@ const getSetting = (key: string): unknown | null => {
  * @param key The string id of the setting for which the value is being retrieved
  * @param newSetting The value that is to be stored. Setting the new value to `null` is the equivalent of deleting the setting
  */
-const setSetting = (key: string, newSetting: unknown | null) => {
+const setSetting = <T>(key: string, newSetting: SettingType<T>) => {
   localStorage.setItem(key, JSON.stringify(newSetting));
-  messageEmitters.forEach((emitter) => {
+  onDidUpdateSettingEmitters.forEach((emitter) => {
     emitter.emit(key);
   });
 };
@@ -33,10 +35,10 @@ const setSetting = (key: string, newSetting: unknown | null) => {
  * @returns Unsubscriber that should be called whenever the subscription should be deleted
  */
 const subscribeToSetting = (key: string, callback: () => void): Unsubscriber => {
-  let emitter = messageEmitters.get(key);
+  let emitter = onDidUpdateSettingEmitters.get(key);
   if (!emitter) {
-    emitter = new PapiEventEmitter<unknown>();
-    messageEmitters.set(key, emitter);
+    emitter = new PapiEventEmitter<string>();
+    onDidUpdateSettingEmitters.set(key, emitter);
   }
   return emitter.subscribe(callback);
 };
